Update "/all" endpoint to provide "payload" as "***" instead of null when detailed messages are not are requested. Fix Redis message ordering when retrieving unassigned and assigned messages. Move common code to MultiQueue interface default implementation. Update tests to cover changes.
